Let our command stack implement IModelAccess and IModelUpdateNotifier
Also register the command stack as a service for both interfaces, so
that the clients of IModelUpdateNotifier actually don't have to know
whether it is implemented by the command stack or anything else.
