+++ client/packages/glsp-sprotty/src/features/command-palette/command-palette.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 EclipseSource
+ * Copyright (c) 2019 EclipseSource
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
