+++ client/packages/glsp-sprotty/src/features/tool-feedback/feedback-action-dispatcher.ts
@@ -10,7 +10,8 @@
  ******************************************************************************/
 import { inject, injectable } from "inversify";
 import { Action, IActionDispatcher, ILogger, SModelRoot, TYPES } from "sprotty/lib";
-import { CommandStackObserver, ObservableCommandStack } from "../../base/command-stack";
+import { CommandStackObserver, IModelUpdateNotifier } from "../../base/command-stack";
+import { GLSP_TYPES } from "../../types";
 
 export interface IFeedbackEmitter { }
 
@@ -46,10 +47,10 @@ export class FeedbackActionDispatcher implements IFeedbackActionDispatcher, Comm
     protected feedbackEmitters: Map<IFeedbackEmitter, Action[]> = new Map;
 
     constructor(
-        @inject(ObservableCommandStack) protected commandStack: ObservableCommandStack,
+        @inject(GLSP_TYPES.IModelUpdateNotifier) protected modelUpdateNotifier: IModelUpdateNotifier,
         @inject(TYPES.IActionDispatcherProvider) protected actionDispatcher: () => Promise<IActionDispatcher>,
         @inject(TYPES.ILogger) protected logger: ILogger) {
-        this.commandStack.registerObserver(this);
+        this.modelUpdateNotifier.registerObserver(this);
     }
 
     registerFeedback(feedbackEmitter: IFeedbackEmitter, actions: Action[]): void {
