+++ client/packages/workflow-sprotty/src/di.config.ts
@@ -11,7 +11,7 @@
 
 import {
     boundsModule, buttonModule, commandPaletteModule, configureModelElement, ConsoleLogger, defaultGLSPModule, defaultModule, DiamondNodeView, ExpandButtonView, //
-    expandModule, exportModule, fadeModule, GLSPGraph, hoverModule, HtmlRoot, HtmlRootView, LogLevel, modelAccessModule, modelHintsModule, modelSourceModule, //
+    expandModule, exportModule, fadeModule, GLSPGraph, hoverModule, HtmlRoot, HtmlRootView, LogLevel, modelHintsModule, modelSourceModule, //
     openModule, overrideViewerOptions, PreRenderedElement, PreRenderedView, RectangularNode, RectangularNodeView, resizeCommandModule, saveModule, SButton, //
     SCompartment, SCompartmentView, SEdge, selectModule, SGraphView, SLabel, SLabelView, SResizeHandle, SRoutingHandle, SRoutingHandleView, toolFeedbackModule, //
     toolManagerModule, TYPES, undoRedoModule, viewportModule
@@ -54,7 +54,7 @@ export default function createContainer(widgetId: string): Container {
     container.load(defaultModule, selectModule, boundsModule, undoRedoModule, viewportModule,
         hoverModule, fadeModule, exportModule, expandModule, openModule, buttonModule, modelSourceModule,
         workflowDiagramModule, saveModule, executeCommandModule, toolManagerModule, toolFeedbackModule, defaultGLSPModule, modelHintsModule,
-        resizeCommandModule, modelAccessModule, commandPaletteModule);
+        resizeCommandModule, commandPaletteModule);
 
     overrideViewerOptions(container, {
         needsClientLayout: true,
